added flag of Malaysia

+++ Sources/CountriesFlags/Countries/UIImage+Greece.swift
@@ -32,8 +32,8 @@ extension UIImage {
         UIRectFill(CGRect(origin: .zero, size: size))
 
         colorWhite.setFill()
-        for index in stride(from: 2, through: 14, by: 4) {
-            UIRectFill(CGRect(origin: CGPoint(x: 0.0, y: size.height / 18.0 * CGFloat(index)),
+        for index in stride(from: 2.0, through: 14.0, by: 4.0) {
+            UIRectFill(CGRect(origin: CGPoint(x: 0.0, y: size.height / 18.0 * index),
                               size: CGSize(width: size.width, height: size.height / 18.0 * 2.0)))
         }
